JournalAI/requirements.txt
Flask==3.0.2
firebase-admin==6.5.0
google-genai==0.7.0
gunicorn==21.2.0
